examples/todo: adds update mutation
